Move NewNetworkAddons from the apis/hco/v1beta1 package to the operands package
