+++ pkg/controller/operands/networkAddons.go
@@ -1,13 +1,18 @@
 package operands
 
 import (
+	networkaddonsshared "github.com/kubevirt/cluster-network-addons-operator/pkg/apis/networkaddonsoperator/shared"
 	networkaddonsv1 "github.com/kubevirt/cluster-network-addons-operator/pkg/apis/networkaddonsoperator/v1"
+	networkaddonsnames "github.com/kubevirt/cluster-network-addons-operator/pkg/names"
+	hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/pkg/apis/hco/v1beta1"
 	"github.com/kubevirt/hyperconverged-cluster-operator/pkg/controller/common"
 	hcoutil "github.com/kubevirt/hyperconverged-cluster-operator/pkg/util"
 	objectreferencesv1 "github.com/openshift/custom-resource-status/objectreferences/v1"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/reference"
+	sdkapi "kubevirt.io/controller-lifecycle-operator-sdk/pkg/sdk/api"
 	"reflect"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -15,7 +20,7 @@ import (
 type cnaHandler genericOperand
 
 func (h cnaHandler) Ensure(req *common.HcoRequest) *EnsureResult {
-	networkAddons := req.Instance.NewNetworkAddons()
+	networkAddons := NewNetworkAddons(req.Instance)
 
 	res := NewEnsureResult(networkAddons)
 	key, err := client.ObjectKeyFromObject(networkAddons)
@@ -87,5 +92,65 @@ func (h cnaHandler) Ensure(req *common.HcoRequest) *EnsureResult {
 	upgradeDone := req.ComponentUpgradeInProgress && isReady && checkComponentVersion(hcoutil.CnaoVersionEnvV, found.Status.ObservedVersion)
 
 	return res.SetUpgradeDone(upgradeDone)
+}
 
+func NewNetworkAddons(hc *hcov1beta1.HyperConverged, opts ...string) *networkaddonsv1.NetworkAddonsConfig {
+
+	cnaoSpec := networkaddonsshared.NetworkAddonsConfigSpec{
+		Multus:      &networkaddonsshared.Multus{},
+		LinuxBridge: &networkaddonsshared.LinuxBridge{},
+		Ovs:         &networkaddonsshared.Ovs{},
+		NMState:     &networkaddonsshared.NMState{},
+		KubeMacPool: &networkaddonsshared.KubeMacPool{},
+	}
+
+	cnaoInfra := hcoConfig2CnaoPlacement(hc.Spec.Infra.NodePlacement)
+	cnaoWorkloads := hcoConfig2CnaoPlacement(hc.Spec.Workloads.NodePlacement)
+	if cnaoInfra != nil || cnaoWorkloads != nil {
+		cnaoSpec.PlacementConfiguration = &networkaddonsshared.PlacementConfiguration{
+			Infra:     cnaoInfra,
+			Workloads: cnaoWorkloads,
+		}
+	}
+
+	return &networkaddonsv1.NetworkAddonsConfig{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      networkaddonsnames.OPERATOR_CONFIG,
+			Labels:    getLabels(hc),
+			Namespace: getNamespace(hcoutil.UndefinedNamespace, opts),
+		},
+		Spec: cnaoSpec,
+	}
+}
+
+func hcoConfig2CnaoPlacement(hcoConf *sdkapi.NodePlacement) *networkaddonsshared.Placement {
+	if hcoConf == nil {
+		return nil
+	}
+	empty := true
+	cnaoPlacement := &networkaddonsshared.Placement{}
+	if hcoConf.Affinity != nil {
+		empty = false
+		hcoConf.Affinity.DeepCopyInto(&cnaoPlacement.Affinity)
+	}
+
+	for _, hcoTol := range hcoConf.Tolerations {
+		empty = false
+		cnaoTol := corev1.Toleration{}
+		hcoTol.DeepCopyInto(&cnaoTol)
+		cnaoPlacement.Tolerations = append(cnaoPlacement.Tolerations, cnaoTol)
+	}
+
+	if len(hcoConf.NodeSelector) > 0 {
+		empty = false
+		cnaoPlacement.NodeSelector = make(map[string]string)
+		for k, v := range hcoConf.NodeSelector {
+			cnaoPlacement.NodeSelector[k] = v
+		}
+	}
+
+	if empty {
+		return nil
+	}
+	return cnaoPlacement
 }
