+++ pkg/controller/operands/networkAddons_test.go
@@ -17,6 +17,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/reference"
+	sdkapi "kubevirt.io/controller-lifecycle-operator-sdk/pkg/sdk/api"
 )
 
 var _ = Describe("CNA Operand", func() {
@@ -31,7 +32,7 @@ var _ = Describe("CNA Operand", func() {
 		})
 
 		It("should create if not present", func() {
-			expectedResource := hco.NewNetworkAddons()
+			expectedResource := NewNetworkAddons(hco)
 			cl := commonTestUtils.InitClient([]runtime.Object{})
 			handler := &cnaHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
 			res := handler.Ensure(req)
@@ -53,7 +54,7 @@ var _ = Describe("CNA Operand", func() {
 		})
 
 		It("should find if present", func() {
-			expectedResource := hco.NewNetworkAddons()
+			expectedResource := NewNetworkAddons(hco)
 			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
 			cl := commonTestUtils.InitClient([]runtime.Object{hco, expectedResource})
 			handler := &cnaHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
@@ -89,7 +90,7 @@ var _ = Describe("CNA Operand", func() {
 		})
 
 		It("should find reconcile to default", func() {
-			existingResource := hco.NewNetworkAddons()
+			existingResource := NewNetworkAddons(hco)
 			existingResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", existingResource.Namespace, existingResource.Name)
 			existingResource.Spec.ImagePullPolicy = corev1.PullAlways // set non-default value
 
@@ -112,7 +113,7 @@ var _ = Describe("CNA Operand", func() {
 		})
 
 		It("should add node placement if missing in CNAO", func() {
-			existingResource := hco.NewNetworkAddons()
+			existingResource := NewNetworkAddons(hco)
 
 			hco.Spec.Infra = hcov1beta1.HyperConvergedConfig{commonTestUtils.NewHyperConvergedConfig()}
 			hco.Spec.Workloads = hcov1beta1.HyperConvergedConfig{commonTestUtils.NewHyperConvergedConfig()}
@@ -149,7 +150,7 @@ var _ = Describe("CNA Operand", func() {
 			hcoNodePlacement := commonTestUtils.NewHco()
 			hcoNodePlacement.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
 			hcoNodePlacement.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
-			existingResource := hcoNodePlacement.NewNetworkAddons()
+			existingResource := NewNetworkAddons(hcoNodePlacement)
 
 			cl := commonTestUtils.InitClient([]runtime.Object{hco, existingResource})
 			handler := &cnaHandler{Client: cl, Scheme: commonTestUtils.GetScheme()}
@@ -175,7 +176,7 @@ var _ = Describe("CNA Operand", func() {
 
 			hco.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
 			hco.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
-			existingResource := hco.NewNetworkAddons()
+			existingResource := NewNetworkAddons(hco)
 
 			// now, modify HCO's node placement
 			seconds3 := int64(3)
@@ -213,7 +214,7 @@ var _ = Describe("CNA Operand", func() {
 		It("should overwrite node placement if directly set on CNAO CR", func() {
 			hco.Spec.Infra = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
 			hco.Spec.Workloads = hcov1beta1.HyperConvergedConfig{NodePlacement: commonTestUtils.NewHyperConvergedConfig()}
-			existingResource := hco.NewNetworkAddons()
+			existingResource := NewNetworkAddons(hco)
 
 			// mock a reconciliation triggered by a change in CNAO CR
 			req.HCOTriggered = false
@@ -259,7 +260,7 @@ var _ = Describe("CNA Operand", func() {
 		})
 
 		It("should handle conditions", func() {
-			expectedResource := hco.NewNetworkAddons()
+			expectedResource := NewNetworkAddons(hco)
 			expectedResource.ObjectMeta.SelfLink = fmt.Sprintf("/apis/v1/namespaces/%s/dummies/%s", expectedResource.Namespace, expectedResource.Name)
 			expectedResource.Status.Conditions = []conditionsv1.Condition{
 				conditionsv1.Condition{
@@ -318,6 +319,135 @@ var _ = Describe("CNA Operand", func() {
 				Reason:  "NetworkAddonsConfigDegraded",
 				Message: "NetworkAddonsConfig is degraded: Bar",
 			}))
+		})
+	})
+
+	Context("hcoConfig2CnaoPlacement", func() {
+		seconds1, seconds2 := int64(1), int64(2)
+		tolr1 := corev1.Toleration{
+			Key: "key1", Operator: "operator1", Value: "value1", Effect: "effect1", TolerationSeconds: &seconds1,
+		}
+		tolr2 := corev1.Toleration{
+			Key: "key2", Operator: "operator2", Value: "value2", Effect: "effect2", TolerationSeconds: &seconds2,
+		}
+
+		It("Should return nil if HCO's input is empty", func() {
+			Expect(hcoConfig2CnaoPlacement(&sdkapi.NodePlacement{})).To(BeNil())
+		})
+
+		It("Should return only NodeSelector", func() {
+			hcoConf := &sdkapi.NodePlacement{
+				NodeSelector: map[string]string{
+					"key1": "value1",
+					"key2": "value2",
+				},
+			}
+			cnaoPlacement := hcoConfig2CnaoPlacement(hcoConf)
+			Expect(cnaoPlacement).ToNot(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector).ToNot(BeNil())
+			Expect(cnaoPlacement.Tolerations).To(BeNil())
+			Expect(cnaoPlacement.Affinity.NodeAffinity).To(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAffinity).To(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAntiAffinity).To(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector["key1"]).Should(Equal("value1"))
+			Expect(cnaoPlacement.NodeSelector["key2"]).Should(Equal("value2"))
+		})
+
+		It("Should return only Tolerations", func() {
+			hcoConf := &sdkapi.NodePlacement{
+				Tolerations: []corev1.Toleration{tolr1, tolr2},
+			}
+			cnaoPlacement := hcoConfig2CnaoPlacement(hcoConf)
+			Expect(cnaoPlacement).ToNot(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector).To(BeNil())
+			Expect(cnaoPlacement.Tolerations).ToNot(BeNil())
+			Expect(cnaoPlacement.Affinity.NodeAffinity).To(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAffinity).To(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAntiAffinity).To(BeNil())
+
+			Expect(cnaoPlacement.Tolerations[0]).Should(Equal(tolr1))
+			Expect(cnaoPlacement.Tolerations[1]).Should(Equal(tolr2))
+		})
+
+		It("Should return only Affinity", func() {
+			affinity := &corev1.Affinity{
+				NodeAffinity: &corev1.NodeAffinity{
+					RequiredDuringSchedulingIgnoredDuringExecution: &corev1.NodeSelector{
+						NodeSelectorTerms: []corev1.NodeSelectorTerm{
+							{
+								MatchExpressions: []corev1.NodeSelectorRequirement{
+									{Key: "key1", Operator: "operator1", Values: []string{"value11, value12"}},
+									{Key: "key2", Operator: "operator2", Values: []string{"value21, value22"}},
+								},
+								MatchFields: []corev1.NodeSelectorRequirement{
+									{Key: "key1", Operator: "operator1", Values: []string{"value11, value12"}},
+									{Key: "key2", Operator: "operator2", Values: []string{"value21, value22"}},
+								},
+							},
+						},
+					},
+				},
+			}
+			hcoConf := &sdkapi.NodePlacement{
+				Affinity: affinity,
+			}
+			cnaoPlacement := hcoConfig2CnaoPlacement(hcoConf)
+			Expect(cnaoPlacement).ToNot(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector).To(BeNil())
+			Expect(cnaoPlacement.Tolerations).To(BeNil())
+			Expect(cnaoPlacement.Affinity.NodeAffinity).ToNot(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAffinity).To(BeNil())
+			Expect(cnaoPlacement.Affinity.PodAntiAffinity).To(BeNil())
+
+			Expect(cnaoPlacement.Affinity.NodeAffinity).Should(Equal(affinity.NodeAffinity))
+		})
+
+		It("Should return the whole object", func() {
+			hcoConf := &sdkapi.NodePlacement{
+
+				NodeSelector: map[string]string{
+					"key1": "value1",
+					"key2": "value2",
+				},
+				Affinity: &corev1.Affinity{
+					NodeAffinity: &corev1.NodeAffinity{
+						RequiredDuringSchedulingIgnoredDuringExecution: &corev1.NodeSelector{
+							NodeSelectorTerms: []corev1.NodeSelectorTerm{
+								{
+									MatchExpressions: []corev1.NodeSelectorRequirement{
+										{Key: "key1", Operator: "operator1", Values: []string{"value11, value12"}},
+										{Key: "key2", Operator: "operator2", Values: []string{"value21, value22"}},
+									},
+									MatchFields: []corev1.NodeSelectorRequirement{
+										{Key: "key1", Operator: "operator1", Values: []string{"value11, value12"}},
+										{Key: "key2", Operator: "operator2", Values: []string{"value21, value22"}},
+									},
+								},
+							},
+						},
+					},
+				},
+				Tolerations: []corev1.Toleration{tolr1, tolr2},
+			}
+
+			cnaoPlacement := hcoConfig2CnaoPlacement(hcoConf)
+			Expect(cnaoPlacement).ToNot(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector).ToNot(BeNil())
+			Expect(cnaoPlacement.Tolerations).ToNot(BeNil())
+			Expect(cnaoPlacement.Affinity.NodeAffinity).ToNot(BeNil())
+
+			Expect(cnaoPlacement.NodeSelector["key1"]).Should(Equal("value1"))
+			Expect(cnaoPlacement.NodeSelector["key2"]).Should(Equal("value2"))
+
+			Expect(cnaoPlacement.Tolerations[0]).Should(Equal(tolr1))
+			Expect(cnaoPlacement.Tolerations[1]).Should(Equal(tolr2))
+
+			Expect(cnaoPlacement.Affinity.NodeAffinity).Should(Equal(hcoConf.Affinity.NodeAffinity))
 		})
 	})
 })
